+++ src/git.rs
@@ -2,9 +2,8 @@ use std::path::Path;
 
 use git2::{Repository, Status};
 
-pub fn is_file_dirty(path: &Path) -> Option<bool> {
+pub fn file_status(path: &Path) -> Option<Status> {
     let repository = Repository::discover(path).ok()?;
     let relative_path = path.strip_prefix(repository.path().parent()?).ok()?;
-    let status = repository.status_file(relative_path).ok()?;
-    Some(status != Status::CURRENT && !status.contains(Status::IGNORED))
+    repository.status_file(relative_path).ok()
 }