+++ CHANGELOG.md
@@ -10,11 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Don't error on uncommitted changes when running with `--check`
+- Add `--allow-staged`
+- Don't check git status when running `--check`
 
 ### Changed
 
-- **Breaking:** Rename `--force` to `--allow-dirty`
+- **Breaking:** Rename `--force` and `-f` to `--allow-dirty`
 
 ## [0.7.0] - 2025-07-17